app/requirements.txt
pdfkit
python-slugify
jinja2
pypdf2